Resolvi todos os erros da base do projeto e criei o get para listar os pagamentos e falta resolver o erro 401 do post

+++ BackEnd/cashhub/src/main/java/com/cashub/cashhubbackend/cashub/dto/PaymentRequest.java
@@ -2,4 +2,4 @@ package com.cashub.cashhubbackend.cashub.dto;
 
 import java.math.BigDecimal;
 
-public record PaymentRequest(String cardNumber, String expirationDate, String cvv, BigDecimal amount, String cardType, String token) {}
+public record PaymentRequest(String cardNumber, String expirationDate, String cvv, Double amount, String cardType, String token) {}
